Explicitly check for files when gzipping.
When gzip is given a directory, it prints a message saying that the file in question is a directory and that it will be ignored. BUT... it seems it also returns exit code "2", and this will cause Jenkins to consider this script a failure and then the whole step fails. SO... I add code to explicitly check that $f is a _regular_ file before calling gzip.